requirements.txt
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_iam
aws-cdk.aws_imagebuilder
aws-cdk.aws_autoscaling
